fix: styled-component 병합 분리되어있던 스타일 컴포넌트를 합쳐주었습니다
